# enabled memmap in tiff reader to improve sino read performance