memmap=True enables memory-mapped file. These files are used for accessing small segments of large files on disk, without reading the entire file into memory. More at http://docs.scipy.org/doc/numpy/reference/generated/numpy.memmap.html 

read test on APS 1-ID tiff files:
read of 1801 x (2kx2k): 15s (memmap=False) vs 9s (memmap=True)
read of 1801 x (100x2k): 11s  (memmap=False) vs 2s  (memmap=True)